requirements.txt
dash==2.14.2
dash-bootstrap-components==1.5.0
dash-leaflet==1.0.4
dash-extensions==1.0.4
pandas==2.2.0
requests==2.31.0
openai==1.12.0
python-dotenv==1.0.1
